% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gridded_data.R
\name{add_date_and_season}
\alias{add_date_and_season}
\title{Add Date and Season to a data table}
\usage{
add_date_and_season(data = D, year = 2001)
}
\arguments{
\item{data}{a data table read from a year of reanalysis data}

\item{year}{the corresponding reanalysis year}
}
\value{
the data with two added rows: date and season
}
\description{
Add Date and Season to a data table
}
