% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gridded_data.R
\name{calculate_descriptives}
\alias{calculate_descriptives}
\title{Calculate descriptives: mean, median, percentiles}
\usage{
calculate_descriptives(data = D, only_mean = FALSE, name = "Quantity")
}
\arguments{
\item{data}{a data table of dimensions (days x monitors).}

\item{only_mean}{logical.  If TRUE, onthe the annual mean is returned.
If FALSE, the mean, median, 1%, 5%, 10%, 90%, 95%, 99%
percentiles are returned.}

\item{name}{the character that will be used a prefix in column names}
}
\value{
a (monitor x number of descriptives) data table
}
\description{
Calculate descriptives: mean, median, percentiles
}
