% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gridded_data.R
\name{create_filename}
\alias{create_filename}
\title{Create filename}
\usage{
create_filename(quantity = "air.sfc", year = 2000)
}
\arguments{
\item{quantity}{the quantity to be processed}

\item{year}{the corresponding reanalysis year}
}
\description{
Create filename
}
