% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{negrid}
\alias{negrid}
\title{Grid cell information for New England}
\format{A data frame with 217660 rows and 3 variables:
\describe{
  \item{SiteCode}{Site code}
  \item{Lat}{Latitude}
  \item{Lon}{Longitude}
}}
\usage{
negrid
}
\description{
A dataset containing the site code, latitude and longitude
of the grid cells in New England
}
\keyword{datasets}
