% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gridded_data.R
\name{process_gridded_data}
\alias{process_gridded_data}
\title{Process gridded data}
\usage{
process_gridded_data(quantity = "air.sfc", year = 2001, seasonal = TRUE,
  only_mean = FALSE, folder = "~/Documents/Tmp/MeasurementError")
}
\arguments{
\item{quantity}{the quantity to be processed}

\item{year}{the corresponding reanalysis year}

\item{seasonal}{logical.}

\item{only_mean}{logical.  If TRUE, onthe the annual mean is returned.
If FALSE, the mean, median, 5%, 10%, 90%, 95%
percentiles are returned.}

\item{folder}{the folder where the reanalysis data is downloaded}
}
\value{
a data table (monitor x number of descriptives)
}
\description{
Process gridded data
}
